Se movió el método ApplicationInitProcess a rviaService y se modificó en su uso para Checkmarx y Aplicaciones. Corrección de los mensajes de error al convertir PDF a CSV.
