+++ src/rvia/rvia.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
+import { BadRequestException, forwardRef, Inject, Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
 import { CreateRviaDto } from './dto/create-rvia.dto';
 import { UpdateRviaDto } from './dto/update-rvia.dto';
 import { ApplicationsService } from 'src/applications/applications.service';
@@ -7,6 +7,7 @@ import { ErrorRVIA } from './helpers/errors-rvia';
 import { CheckmarxService } from 'src/checkmarx/checkmarx.service';
 import { ApplicationstatusService } from 'src/applicationstatus/applicationstatus.service';
 import { ConfigService } from '@nestjs/config';
+import { Application } from 'src/applications/entities/application.entity';
 
 const addon = require(process.env.RVIA_PATH);
 
@@ -16,9 +17,10 @@ export class RviaService {
   private readonly crviaEnvironment: number;
 
   constructor(
-
+    @Inject(forwardRef(() => ApplicationsService))
     private readonly applicationService: ApplicationsService,
     private readonly encryptionService: CommonService,
+    @Inject(forwardRef(() => CheckmarxService))
     private readonly checkmarxService: CheckmarxService,
     private readonly configService: ConfigService,
   ) {
@@ -88,6 +90,36 @@ export class RviaService {
     const responseConvert = { ...aplicacion, ...{ isProccessValid,  message} };
 
     return responseConvert;
+  }
+
+  ApplicationInitProcess(aplicacion:Application, obj: any){
+    // Base de datos: 1 = Producción 2 = Desarrollo
+    // const obj = new addon.CRvia(2);
+    var isValidProcess = true;
+    var messageRVIA;
+    //  -------------------------------- Parámetros de Entrada --------------------------------
+    const lID = aplicacion.idu_proyecto;
+    const lEmployee = aplicacion.user.numero_empleado;
+    const ruta_proyecto = this.encryptionService.decrypt(aplicacion.sourcecode.nom_directorio);
+    const tipo_proyecto = aplicacion.num_accion;
+    const iConIA = 1;
+    // const Bd = 1 = Producion 2 = Desarrollo
+  
+    const bConDoc   = Array.isArray(aplicacion.opc_arquitectura) && aplicacion.opc_arquitectura.length > 1 ? aplicacion.opc_arquitectura[1] : false;
+    const bConCod   = Array.isArray(aplicacion.opc_arquitectura) && aplicacion.opc_arquitectura.length > 2 ? aplicacion.opc_arquitectura[2] : false;
+    const bConTest  = Array.isArray(aplicacion.opc_arquitectura) && aplicacion.opc_arquitectura.length > 3 ? aplicacion.opc_arquitectura[3] : false;
+    const bCalifica = Array.isArray(aplicacion.opc_arquitectura) && aplicacion.opc_arquitectura.length > 4 ? aplicacion.opc_arquitectura[4] : false;
+
+    const initProcessResult = obj.initProcess( lID, lEmployee, ruta_proyecto, tipo_proyecto, iConIA, bConDoc, bConCod, bConTest, bCalifica);
+    
+    if( initProcessResult == 1){
+      messageRVIA = "Proceso IA Iniciado Correctamente";
+    }else{
+      isValidProcess = false;
+      messageRVIA = ErrorRVIA[initProcessResult];
+    }
+
+    return { isValidProcess, messageRVIA };
   }
 
   getVersion() {
